Keep RTU server running
Sometimes the bus may receive incomplete or bad RTU Frames which cause the listening loop to exit. This is not ideal especially for remote devices. Simply jump back to the top of the loop and wait for the next frame. This discards the bad frame and awaits for the next frame, ensuring RTU-server runs without terminating or processing bad frames.